fix : lint修正、責任分離
