+++ front/src/app/(authorized)/binder/[cardId]/edit/page.tsx
@@ -1,20 +1,29 @@
-import { notFound } from "next/navigation";
-import { getCardFromServer } from "@/lib/server";
-import CardEditContainer from "@/components/binder/CardEditContainer";
+import { notFound, redirect } from "next/navigation";
+import EditCardForm from "@/components/binder/EditCardForm";
+import { getCardFromServer, getSession } from "@/lib/server";
 
 interface Props {
-    params: Promise<{
-        cardId: string;
-    }>;
+  params: Promise<{
+    cardId: string;
+  }>;
 }
 
 export default async function EditCardPage({ params }: Props) {
-    const { cardId } = await params;
-    const card = await getCardFromServer(cardId);
+  const { cardId } = await params;
+  const [card, sessionUser] = await Promise.all([
+    getCardFromServer(cardId),
+    getSession(),
+  ]);
 
-    if (!card) {
-        notFound();
-    }
+  if (!card) {
+    notFound();
+  }
 
-    return <CardEditContainer card={card} />;
+  // Server-side Access Control
+  // If not logged in or not creator, redirect to details page
+  if (!sessionUser || sessionUser.uid !== card.creatorId) {
+    redirect(`/binder/${cardId}`);
+  }
+
+  return <EditCardForm card={card} />;
 }
